Testing and changing code

+++ HR_CMS.Data/DataClasses/Contact.cs
@@ -18,7 +18,7 @@ namespace HR_CMS.Data
         public int ContactId { get; set; }
 
         [Required]
-        public int PhoneNumber { get; set; }
+        public double PhoneNumber { get; set; }
 
         [Required]
         public string Email { get; set; }
