+++ HR_CMS.Models/02 Contact/ContactDetail.cs
@@ -15,7 +15,7 @@ namespace HR_CMS.Models
 
         public int ContactId { get; set; }
 
-        public int PhoneNumber { get; set; }
+        public double PhoneNumber { get; set; }
 
         public string Email { get; set; }
